Simply declarative events.
This patch simplifies declarative events such that there are only two notions:

1) A controller: This is found by walking up the DOM until a shadow host or lightDom controller (i.e. polymer-body) is encountered.

2) A method name: When event is handled by the controller, it attempts to find a member function with a matching name and invoke it.

Notably, this removes the behavior (that apparently no one was actually using), where by a mustache could contain a *path* (e.g. foo.bar) and the function foo.bar is looked up, not from the model, from the controller and invoked with the controller as the receiver value.

R=sorvell
BUG=

Review URL: https://codereview.appspot.com/97750045