Support Bower's main-files feature
Bower packages can provide a list of files that are important to
projects using this package. Some build tools use this list to copy,
concatenate, or munge the contents of the files in some way without
touching the files in bower_components.

This commit adds the Polymer files to the list of files that should be
delivered to a build tool.