Add expose method to be able to delegate Polymer object functions.
When passing callbacks to external APIs, it is common to fetch
the `this` reference from the callback. This `this` references the
element that invoked the callback.

However, in order to be able to use Polymer functions, the `this`
that originally referenced the Polymer object, is lost.

Therefore in order to maintain both references, expose links `this`
to the Polymer object and the first argument references the element
that called the callback.