Add support for properties defined in a behavior.
For correctness this is implemented using `getPropertyInfo()` though
this is not optimized at all. Note that `_propertyInfo` is not ready
when we parse the annotations (which is the earliest call to
`_parseMethod` right now). There are several options to make this right,
but these are not part of this feature PR.