Convert object to class for better compilation
Under some closure flags the hostStack object was being split out into three unrelated functions. Converting it to a class makes the code more regular, and compile with more optimizations on.

There's almost definitely a way to express what we were doing with closure annotations, but this seems more robust to me.